New features for the 2024 compiler construction course

+++ copra/src/phase_driver.c
@@ -143,8 +143,8 @@ static void SkipPhase(struct ccn_phase *phase) {
 
 char *Checkbreakpoint(char *name)
 {
-    int len = STRlen(name);
-    for (int i = 0; i < len; i++) {
+    size_t len = STRlen(name);
+    for (size_t i = 0; i < len; i++) {
         if (phase_driver.breakpoint[i] == '\0') {
             return NULL;
         }
